+++ assets/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 - feat: add troubleshoot page
 - feat: cleanup scan UI by hiding multiple network interfaces inside the scan button
-- feat(android): share apk
+- feat(android): share apk and install apk
 - feat(android): Android TV support
 - feat(windows): left click on tray icon opens app
 - feat(linux): add Control+Q shortcut to exit app